+++ examples/.joyride/scripts/webview/example.cljs
@@ -1,19 +1,18 @@
 (ns webview.example
   (:require ["path" :as path]
             ["vscode" :as vscode]
-            [joyride.core :as joyride]
-            [promesa.core :as p]))
+            [joyride.core :as joyride]))
 
-(defn main []
-  (p/let [panel (vscode/window.createWebviewPanel
-                 "My webview!" "Scittle"
-                 vscode/ViewColumn.One
-                 {:enableScripts true})
-          uri (vscode/Uri.file (path/join vscode/workspace.rootPath
-                                          ".joyride" "scripts" "webview"
-                                          "page.html"))
-          data (vscode/workspace.fs.readFile uri)
-          html (.decode (js/TextDecoder. "utf-8") data)]
+(defn ^:async main []
+  (let [panel (vscode/window.createWebviewPanel
+               "My webview!" "Scittle"
+               vscode/ViewColumn.One
+               {:enableScripts true})
+        uri (vscode/Uri.file (path/join vscode/workspace.rootPath
+                                        ".joyride" "scripts" "webview"
+                                        "page.html"))
+        data (await (vscode/workspace.fs.readFile uri))
+        html (.decode (js/TextDecoder. "utf-8") data)]
     (set! (.. panel -webview -html) (str html))))
 
 (when (= (joyride/invoked-script) joyride/*file*)