+++ CHANGELOG.md
@@ -4,11 +4,10 @@ Changes to Joyride
 
 ## [Unreleased]
 
-- Update SCI to a version that supports async/await
+- Update SCI to latest version, to get async/await for Joyride scripts
 - Add `who` badge and rework info line in terminal output
 - Log agent evaluations to Calva's output subscriber bus when Calva is installed
 - Update bundled agent instructions
-- Make most examples use SCI await instead of Promesa
 
 ## [0.0.72] - 2025-12-20